# BACKEND/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API pour OC elites du congo",
        "description": "Oc elites du congo",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://elite.com/api"
        }
    ],
    "paths": {
...
        "/admin/players/store": {
            "post": {
                "tags": [
                    "ADMINISTRATION__JOUEURS"
                ],
                "summary": "Ajouter un joueur",
                "operationId": "1c33338fe7761ba07553a823d3652e31",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Mulamba"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Enock"
                                    },
                                    "postnom": {
                                        "type": "string",
                                        "example": "Kabamba"
                                    },
                                    "date_naissance": {
                                        "type": "date",
                                        "example": "1998-02-18"
                                    },
                                    "lieu_naissance": {
                                        "type": "string",
                                        "example": "Mbuji-mayi"
                                    },
                                    "date_debut_carriere": {
                                        "type": "date",
                                        "example": "2010-08-10"
                                    },
                                    "date_signature": {
                                        "type": "date",
                                        "example": "2020-09-04"
                                    },
                                    "nationalite": {
                                        "type": "string",
                                        "example": "Congolaise"
                                    },
                                    "poids": {
                                        "type": "float",
                                        "example": 70
                                    },
                                    "taille": {
                                        "type": "float",
                                        "example": 189
                                    },
                                    "dorsale": {
                                        "type": "integer",
                                        "example": 9
                                    },
                                    "position": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "descriptions": {
                                        "type": "string",
                                        "example": "Une description"
                                    },
                                    "cover": {
                                        "type": "file"
                                    },
                                    "images[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/players/show/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__JOUEURS"
                ],
                "summary": "Renvoie les informations d'un joueur par son ID",
                "operationId": "06f5d913416127b27827a677e8c7d88a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/players/update": {
            "post": {
                "tags": [
                    "ADMINISTRATION__JOUEURS"
                ],
                "summary": "Editer les informations d'un Joueur",
                "operationId": "5b6f22ece3aed3b6751095913948c326",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nom": {
                                        "type": "string",
                                        "example": "Mulamba"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Enock"
                                    },
                                    "postnom": {
                                        "type": "string",
                                        "example": "Kabamba"
                                    },
                                    "date_naissance": {
                                        "type": "date",
                                        "example": "1998-02-18"
                                    },
                                    "lieu_naissance": {
                                        "type": "string",
                                        "example": "Mbuji-mayi"
                                    },
                                    "date_debut_carriere": {
                                        "type": "date",
                                        "example": "2010-08-10"
                                    },
                                    "date_signature": {
                                        "type": "date",
                                        "example": "2020-09-04"
                                    },
                                    "nationalite": {
                                        "type": "string",
                                        "example": "Congolaise"
                                    },
                                    "poids": {
                                        "type": "float",
                                        "example": 70
                                    },
                                    "taille": {
                                        "type": "float",
                                        "example": 70
                                    },
                                    "dorsale": {
                                        "type": "integer",
                                        "example": 9
                                    },
                                    "position": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "descriptions": {
                                        "type": "string",
                                        "example": "Une modif de la description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/players/update/image": {
            "post": {
                "tags": [
                    "ADMINISTRATION__JOUEURS"
                ],
                "summary": "Moofier une photo liée à un évènement",
                "operationId": "6901d3b0c918789d733075ba5a07a9eb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "player_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "image": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/players/delete/image": {
            "post": {
                "tags": [
                    "ADMINISTRATION__JOUEURS"
                ],
                "summary": "Supprimer une photo liée à un évènement",
                "operationId": "62e6d4d0d10a26c6f37168c6a44da186",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": [
            []
        ]
    }
}